🗃️ Align schema with DB, also pushed all missing indexes to prod for performance boost

+++ app/drizzle/schema.ts
@@ -1808,7 +1808,7 @@ export const gameSetting = mysqlTable(
     id: varchar("id", { length: 191 }).primaryKey().notNull(),
     name: varchar("name", { length: 191 }).notNull(),
     time: datetime("time", { mode: "date", fsp: 3 }).notNull(),
-    value: tinyint("value").default(0).notNull(),
+    value: int("value").default(0).notNull(),
   },
   (table) => {
     return { name: index("name").on(table.name) };
